refactor: modularize large files into focused components
Breaking down monolithic files into smaller, more maintainable modules:

## apiValidationService.ts (576 → 3 lines)
- Split validation logic by provider (OpenRouter, Gateway, Local, OpenAI, Anthropic, Google)
- New structure: apiValidation/{index,types,providers/*}
- Each provider now has its own focused validation module

## mcpHandlers.ts (725 → 2 lines)
- Organized 30+ IPC handlers by functionality
- New structure: mcpHandlers/{index,connection,configuration,tools,health,extraction,registry}
- Clear separation between connection management, config, tools, and health monitoring

## mcpService.ts (674 → 3 lines)
- Separated transport logic, command resolution, diagnostics, and registry
- New structure: mcp/{index,types,transports,commandResolver,diagnostics,registry}
- Each module handles a specific aspect of MCP service functionality

## preload.ts (617 → 247 lines)
- Split API definitions by domain (app, chat, models, database, mcp, etc.)
- New structure: preload/{types/,api/*}
- 12 focused API modules replacing single monolithic file

## Benefits
✅ Improved maintainability - smaller, focused files
✅ Better scalability - easy to add new providers/handlers
✅ Enhanced readability - clear module responsibilities
✅ Zero breaking changes - all original files now re-export modules
✅ TypeScript compilation passes without errors

Total: 29 new modules created, ~2,500 lines refactored
