+++ src/main/services/apiValidation/types.ts
@@ -1,0 +1,6 @@
+export type { ValidationResult, ProviderValidationConfig } from '../../../types/wizard';
+
+export interface ModelsResponse {
+  data?: any[];
+  models?: any[];
+}
