+++ src/preload/api/app.ts
@@ -1,0 +1,33 @@
+import { ipcRenderer } from 'electron';
+import type { DeepLinkAction } from '../types';
+
+export const appApi = {
+  getVersion: () => ipcRenderer.invoke('levante/app/version'),
+  getPlatform: () => ipcRenderer.invoke('levante/app/platform'),
+  getSystemTheme: () => ipcRenderer.invoke('levante/app/theme'),
+  onSystemThemeChanged: (callback: (theme: { shouldUseDarkColors: boolean; themeSource: string }) => void) => {
+    const listener = (_event: any, theme: { shouldUseDarkColors: boolean; themeSource: string }) => {
+      callback(theme);
+    };
+    ipcRenderer.on('levante/app/theme-changed', listener);
+
+    // Return cleanup function
+    return () => {
+      ipcRenderer.removeListener('levante/app/theme-changed', listener);
+    };
+  },
+
+  checkForUpdates: () => ipcRenderer.invoke('levante/app/check-for-updates'),
+
+  onDeepLink: (callback: (action: DeepLinkAction) => void) => {
+    const listener = (_event: any, action: DeepLinkAction) => {
+      callback(action);
+    };
+    ipcRenderer.on('levante/deep-link/action', listener);
+
+    // Return cleanup function
+    return () => {
+      ipcRenderer.removeListener('levante/deep-link/action', listener);
+    };
+  },
+};
